refactor: modulizate left components
