+++ src/components/draw/drawon.ts
@@ -1,0 +1,9 @@
+import { FrameBuffer, GameObj } from "@/types";
+
+export function drawon(c: FrameBuffer) {
+    return {
+        add(this: GameObj) {
+            this.canvas = c;
+        },
+    };
+}
