+++ src/components/draw/mask.ts
@@ -1,0 +1,8 @@
+import { Mask, MaskComp } from "@/types";
+
+export function mask(m: Mask = "intersect"): MaskComp {
+    return {
+        id: "mask",
+        mask: m,
+    };
+}
