+++ src/components/draw/outline.ts
@@ -1,0 +1,15 @@
+import { rgb } from "@/math";
+import { Color, OutlineComp } from "@/types";
+
+export function outline(
+    width: number = 1,
+    color: Color = rgb(0, 0, 0),
+): OutlineComp {
+    return {
+        id: "outline",
+        outline: {
+            width,
+            color,
+        },
+    };
+}
