+++ src/components/draw/shader.ts
@@ -1,0 +1,21 @@
+import { ShaderComp, Uniform } from "@/types";
+
+export function shader(
+    id: string,
+    uniform?: Uniform | (() => Uniform),
+): ShaderComp {
+    return {
+        id: "shader",
+        shader: id,
+        ...(typeof uniform === "function"
+            ? {
+                uniform: uniform(),
+                update() {
+                    this.uniform = uniform();
+                },
+            }
+            : {
+                uniform: uniform,
+            }),
+    };
+}
