+++ src/components/index.ts
@@ -1,10 +1,32 @@
 export * from "./draw/circle";
+export * from "./draw/drawon";
+export * from "./draw/fadeIn";
+export * from "./draw/mask";
 export * from "./draw/opacity";
+export * from "./draw/outline";
+export * from "./draw/polygon";
+export * from "./draw/raycast";
+export * from "./draw/rect";
+export * from "./draw/shader";
 export * from "./draw/sprite";
 export * from "./draw/text";
+export * from "./draw/uvquad";
+export * from "./level/agent";
+export * from "./level/tile";
 export * from "./misc/health";
+export * from "./misc/lifespan";
+export * from "./misc/state";
+export * from "./misc/stay";
+export * from "./misc/timer";
 export * from "./physics/area";
+export * from "./physics/body";
+export * from "./physics/doubleJump";
 export * from "./transform/anchor";
+export * from "./transform/fixed";
+export * from "./transform/follow";
+export * from "./transform/move";
+export * from "./transform/offscreen";
 export * from "./transform/pos";
 export * from "./transform/rotate";
 export * from "./transform/scale";
+export * from "./transform/z";
