+++ src/components/misc/stay.ts
@@ -1,0 +1,9 @@
+import { StayComp } from "@/types";
+
+export function stay(scenesToStay?: string[]): StayComp {
+    return {
+        id: "stay",
+        stay: true,
+        scenesToStay: scenesToStay,
+    };
+}
