+++ src/components/transform/fixed.ts
@@ -1,0 +1,8 @@
+import { FixedComp } from "@/types";
+
+export function fixed(): FixedComp {
+    return {
+        id: "fixed",
+        fixed: true,
+    };
+}
